requirements.txt
opencv-python
pillow
pyyaml
requests
numpy
